@@ -33,36 +35,36 @@
     "svelte": "^3.55.1"
   },
   "devDependencies": {
-    "@playwright/test": "^1.30.0",
-    "@sveltejs/adapter-vercel": "^2.0.4",
-    "@sveltejs/kit": "^1.7.2",
-    "@sveltejs/package": "2.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.52.0",
-    "@typescript-eslint/parser": "^5.52.0",
-    "autoprefixer": "^10.4.13",
+    "@playwright/test": "^1.32.3",
+    "@sveltejs/adapter-vercel": "^2.4.2",
+    "@sveltejs/kit": "^1.15.8",
+    "@sveltejs/package": "2.0.2",
+    "@typescript-eslint/eslint-plugin": "^5.59.1",
+    "@typescript-eslint/parser": "^5.59.1",
+    "autoprefixer": "^10.4.14",
     "createprops": "^0.4.9",
-    "esbuild": "0.17.9",
-    "eslint": "^8.34.0",
-    "eslint-config-prettier": "^8.6.0",
+    "esbuild": "0.17.18",
+    "eslint": "^8.39.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-svelte3": "^4.0.0",
     "mdsvex": "^0.10.6",
     "mdsvexamples": "^0.3.3",
-    "postcss": "^8.4.21",
+    "postcss": "^8.4.23",
     "postcss-load-config": "^4.0.1",
-    "prettier": "^2.8.4",
-    "prettier-plugin-svelte": "^2.9.0",
+    "prettier": "^2.8.8",
+    "prettier-plugin-svelte": "^2.10.0",
     "prism-themes": "^1.9.0",
-    "publint": "^0.1.9",
-    "svelte": "^3.55.1",
-    "svelte-check": "^3.0.3",
-    "svelte-meta-tags": "^2.6.5",
-    "svelte-preprocess": "^5.0.1",
-    "svelte2tsx": "^0.6.1",
-    "tailwindcss": "^3.2.7",
+    "publint": "^0.1.11",
+    "svelte": "^3.58.0",
+    "svelte-check": "^3.2.0",
+    "svelte-meta-tags": "^2.7.0",
+    "svelte-preprocess": "^5.0.3",
+    "svelte2tsx": "^0.6.11",
+    "tailwindcss": "^3.3.1",
     "tslib": "^2.5.0",
-    "typescript": "^4.9.5",
-    "vite": "^4.1.2",
-    "vitest": "^0.28.5"
+    "typescript": "^5.0.4",
+    "vite": "^4.3.2",
+    "vitest": "^0.30.1"
   },
   "type": "module",
   "keywords": [
