@@ -3,7 +3,8 @@
   import classNames from 'classnames';
   import CloseButton from '../utils/CloseButton.svelte';
 
-  export let color: 'blue' | 'dark' | 'red' | 'green' | 'yellow' | 'indigo' | 'purple' | 'pink' = 'blue';
+  export let color: 'blue' | 'dark' | 'red' | 'green' | 'yellow' | 'indigo' | 'purple' | 'pink' | 'none' =
+    'blue';
   export let large: boolean = false;
   export let border: boolean = false;
   export let href: string | undefined = undefined;
@@ -27,7 +28,8 @@
     ['!yellow']: 'bg-yellow-300 text-yellow-800',
     ['!indigo']: 'bg-indigo-500 text-indigo-100',
     ['!purple']: 'bg-purple-500 text-purple-100',
-    ['!pink']: 'bg-pink-500 text-pink-100'
+    ['!pink']: 'bg-pink-500 text-pink-100',
+    none: ''
   };
 
   const borderedColors = {
@@ -38,7 +40,8 @@
     yellow: 'bg-yellow-100 text-yellow-800 dark:bg-gray-700 dark:text-yellow-300 border-yellow-300',
     indigo: 'bg-indigo-100 text-indigo-800 dark:bg-gray-700 dark:text-indigo-400 border-indigo-400',
     purple: 'bg-purple-100 text-purple-800 dark:bg-gray-700 dark:text-purple-400 border-purple-400',
-    pink: 'bg-pink-100 text-pink-800 dark:bg-gray-700 dark:text-pink-400 border-pink-400'
+    pink: 'bg-pink-100 text-pink-800 dark:bg-gray-700 dark:text-pink-400 border-pink-400',
+    none: ''
   };
 
   const hoverColors = {
@@ -49,7 +52,8 @@
     yellow: 'hover:bg-yellow-200',
     indigo: 'hover:bg-indigo-200',
     purple: 'hover:bg-purple-200',
-    pink: 'hover:bg-pink-200'
+    pink: 'hover:bg-pink-200',
+    none: ''
   };
 
   const baseClass: string = 'font-medium inline-flex items-center justify-center px-2.5 py-0.5';
@@ -69,7 +73,8 @@
     $$props.class
   );
 
-  const closeBtnBaseClass: string = 'inline-flex items-center !p-0.5 !m-0 !ml-2 text-sm bg-transparent rounded-sm focus:!ring-0';
+  const closeBtnBaseClass: string =
+    'inline-flex items-center !p-0.5 !m-0 !ml-2 text-sm bg-transparent rounded-sm focus:!ring-0';
 
   // uncomment the following lines to completely overwrite <CloseButton>'s color palette
   // in case <CloseButton>'s color palette is changed in the future
@@ -93,14 +98,12 @@
     yellow: 'text-yellow-400 hover:text-yellow-900 dark:hover:!bg-yellow-800 dark:hover:text-yellow-300',
     indigo: 'text-indigo-400 hover:text-indigo-900 dark:hover:!bg-indigo-800 dark:hover:text-indigo-300',
     purple: 'text-purple-400 hover:text-purple-900 dark:hover:!bg-purple-800 dark:hover:text-purple-300',
-    pink: 'text-pink-400 hover:text-pink-900 dark:hover:!bg-pink-800 dark:hover:text-pink-300'
+    pink: 'text-pink-400 hover:text-pink-900 dark:hover:!bg-pink-800 dark:hover:text-pink-300',
+    none: ''
   };
 
   let closeBtnClass: string;
-  $: closeBtnClass = classNames(
-    closeBtnBaseClass,
-    closeBtnColors[color]
-  );
+  $: closeBtnClass = classNames(closeBtnBaseClass, closeBtnColors[color]);
 
   let hidden = false;
   const dispatch = createEventDispatcher();
@@ -111,8 +114,8 @@
       hidden = true;
     }, 300);
     dispatch('dismiss', {
-			message: 'The badge will be dismissed.'
-		});
+      message: 'The badge will be dismissed.'
+    });
   };
 </script>
 
@@ -120,7 +123,12 @@
   <slot />
   {#if dismissable}
     <slot name="closeBtn" {handleHide}>
-      <CloseButton {color} on:click={handleHide} size={large ? 'sm' : 'xs'} name="Remove badge" class={closeBtnClass} />
+      <CloseButton
+        {color}
+        on:click={handleHide}
+        size={large ? 'sm' : 'xs'}
+        name="Remove badge"
+        class={closeBtnClass} />
     </slot>
   {/if}
 </svelte:element>
