@@ -5,5 +5,21 @@
 </script>
 
 <Input let:props {...$$restProps}>
-  <input {...props} type="number" bind:value />
+  <input
+    {...props}
+    type="number"
+    bind:value
+    on:blur
+    on:change
+    on:click
+    on:contextmenu
+    on:focus
+    on:keydown
+    on:keypress
+    on:keyup
+    on:mouseover
+    on:mouseenter
+    on:mouseleave
+    on:paste
+    on:input />
 </Input>
