@@ -1,17 +1,19 @@
 <script lang="ts">
   import classNames from 'classnames';
+  import { getContext } from 'svelte';
+  import type { NavbarLiType } from './NavUl.svelte';
 
   export let href: string = '';
   export let active: boolean = false;
-  export let activeClass: string =
-    'text-white bg-blue-700 md:bg-transparent md:text-blue-700 md:dark:text-white dark:bg-blue-600 md:dark:bg-transparent';
-  export let nonActiveClass: string =
-    'text-gray-700 hover:bg-gray-100 md:hover:bg-transparent md:border-0 md:hover:text-blue-700 dark:text-gray-400 md:dark:hover:text-white dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent';
+  export let activeClass: string | undefined = undefined;
+  export let nonActiveClass: string | undefined = undefined;
+
+  const context = getContext<NavbarLiType>('navbar') ?? {};
 
   let liClass: string;
   $: liClass = classNames(
     'block py-2 pr-4 pl-3 md:p-0 rounded md:border-0',
-    active ? activeClass : nonActiveClass,
+    active ? activeClass ?? context.activeClass : nonActiveClass ?? context.nonActiveClass,
     $$props.class
   );
 </script>
