@@ -9,6 +9,7 @@
 </script>
 
 <script lang="ts">
+  import { getContext } from 'svelte';
   import classNames from 'classnames';
   import Button from '$lib/buttons/Button.svelte';
   import Popper from '$lib/utils/Popper.svelte';
@@ -22,6 +23,11 @@
   export let trigger: 'hover' | 'click' = 'hover';
   export let textOutside: boolean = false;
   export let id: string = generateId();
+  export let shadow: 'blue' | 'green' | 'cyan' | 'teal' | 'lime' | 'red' | 'pink' | 'purple' | null = null;
+  export let outline: boolean = false;
+  const group = getContext('group');
+  export let color: 'alternative'| 'blue'| 'cyan'| 'dark'| 'light'| 'lime'| 'green'| 'pink'| 'primary'| 'red'| 'teal'| 'yellow'| 'purple'| 'purpleToBlue'| 'cyanToBlue'| 'greenToBlue'| 'purpleToPink'| 'pinkToOrange'| 'tealToLime'| 'redToYellow' = group ? (outline ? 'dark' : 'alternative') : 'blue';
+  export let gradient: boolean = false;
 
   setContext<SpeedCtxType>('speed-dial', { pill, tooltip, textOutside });
 
@@ -36,7 +42,7 @@
 </script>
 
 <div class={divClass}>
-  <Button {pill} name="Open actions menu" aria-controls={id} aria-expanded="false" color="blue" class="!p-3">
+  <Button {pill} name="Open actions menu" aria-controls={id} aria-expanded="false" {color} {gradient} {outline} {shadow} class="!p-3">
     <slot name="icon">
       <svg
         aria-hidden="true"
