@@ -17,7 +17,8 @@
     | 'blue' = 'default';
   export let name: string | undefined = undefined;
   export let ariaLabel: string | undefined = undefined;
-  export let size: 'xs' | 'sm' | 'md' = 'md';
+  export let size: 'xs' | 'sm' | 'md' | 'lg' = 'md';
+  export let href: string | undefined = undefined;
 
   const colors = {
     dark: 'text-gray-500 hover:text-gray-900 hover:bg-gray-200 dark:text-gray-400 dark:hover:text-white dark:hover:bg-gray-700',
@@ -35,7 +36,8 @@
   const sizing = {
     xs: 'm-0.5 rounded focus:ring-1 p-0.5',
     sm: 'm-0.5 rounded focus:ring-1 p-0.5',
-    md: 'rounded-lg focus:ring-2 p-1.5'
+    md: 'm-0.5 rounded-lg focus:ring-2 p-1.5',
+    lg: 'm-0.5 rounded-lg focus:ring-2 p-2.5'
   };
 
   let buttonClass: string;
@@ -51,11 +53,19 @@
   const svgSizes = {
     xs: 'w-3 h-3',
     sm: 'w-3.5 h-3.5',
-    md: 'w-5 h-5'
+    md: 'w-5 h-5',
+    lg: 'w-5 h-5'
   };
 </script>
 
-<button on:click type="button" {...$$restProps} class={buttonClass} aria-label={ariaLabel ?? name}>
-  {#if name}<span class="sr-only">{name}</span>{/if}
-  <slot svgSize={svgSizes[size]} />
-</button>
+{#if href}
+  <a {href} {...$$restProps} class={buttonClass} aria-label={ariaLabel ?? name}>
+    {#if name}<span class="sr-only">{name}</span>{/if}
+    <slot svgSize={svgSizes[size]} />
+  </a>
+{:else}
+  <button on:click type="button" {...$$restProps} class={buttonClass} aria-label={ariaLabel ?? name}>
+    {#if name}<span class="sr-only">{name}</span>{/if}
+    <slot svgSize={svgSizes[size]} />
+  </button>
+{/if}
