@@ -1,23 +1,23 @@
 <script lang="ts">
-	export let size = '24';
+	export let size = '20';
 	export let color = 'currentColor';
 	export let variation: 'solid' | 'outline' = 'outline';
 	let viewBox: string;
 	let svgpath: string;
-	let svgoutline = `<path d="M19.5 8.25L12 15.75L4.5 8.25" stroke="${color}" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/> `;
-	let svgsolid = `<path fill-rule="evenodd" clip-rule="evenodd" d="M12.5303 16.2803C12.2374 16.5732 11.7626 16.5732 11.4697 16.2803L3.96967 8.78033C3.67678 8.48744 3.67678 8.01256 3.96967 7.71967C4.26256 7.42678 4.73744 7.42678 5.03033 7.71967L12 14.6893L18.9697 7.71967C19.2626 7.42678 19.7374 7.42678 20.0303 7.71967C20.3232 8.01256 20.3232 8.48744 20.0303 8.78033L12.5303 16.2803Z" fill="${color}"/> `;
+	let svgoutline = `<path stroke-linecap="round" stroke-linejoin="round" d="M19 9l-7 7-7-7" stroke="${color}"></path>`;
+	let svgsolid = `<path clip-rule="evenodd" fill-rule="evenodd" d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z" fill="${color}"></path>`;
 	$: switch (variation) {
 		case 'outline':
 			svgpath = svgoutline;
-			viewBox = '0 0 24 24';
+			viewBox = '0 0 20 20';
 			break;
 		case 'solid':
 			svgpath = svgsolid;
-			viewBox = '0 0 24 24';
+			viewBox = '0 0 20 20';
 			break;
 		default:
 			svgpath = svgoutline;
-			viewBox = '0 0 24 24';
+			viewBox = '0 0 20 20';
 	}
 	export let ariaLabel = 'chevron down';
 </script>
