@@ -1,23 +1,23 @@
 <script lang="ts">
-	export let size = '24';
+	export let size = '20';
 	export let color = 'currentColor';
 	export let variation: 'solid' | 'outline' = 'outline';
 	let viewBox: string;
 	let svgpath: string;
-	let svgoutline = `<path d="M15.75 19.5L8.25 12L15.75 4.5" stroke="${color}" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/> `;
-	let svgsolid = `<path fill-rule="evenodd" clip-rule="evenodd" d="M7.71967 12.5303C7.42678 12.2374 7.42678 11.7626 7.71967 11.4697L15.2197 3.96967C15.5126 3.67678 15.9874 3.67678 16.2803 3.96967C16.5732 4.26256 16.5732 4.73744 16.2803 5.03033L9.31066 12L16.2803 18.9697C16.5732 19.2626 16.5732 19.7374 16.2803 20.0303C15.9874 20.3232 15.5126 20.3232 15.2197 20.0303L7.71967 12.5303Z" fill="${color}"/> `;
+	let svgoutline = `<path clip-rule="evenodd" fill-rule="evenodd" d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z" stroke="${color}"></path>`;
+	let svgsolid = `<path clip-rule="evenodd" fill-rule="evenodd" d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z" fill="${color}"></path>`;
 	$: switch (variation) {
 		case 'outline':
 			svgpath = svgoutline;
-			viewBox = '0 0 24 24';
+			viewBox = '0 0 20 20';
 			break;
 		case 'solid':
 			svgpath = svgsolid;
-			viewBox = '0 0 24 24';
+			viewBox = '0 0 20 20';
 			break;
 		default:
 			svgpath = svgoutline;
-			viewBox = '0 0 24 24';
+			viewBox = '0 0 20 20';
 	}
 	export let ariaLabel = 'chevron left';
 </script>
