@@ -1,23 +1,23 @@
 <script lang="ts">
-	export let size = '24';
+	export let size = '20';
 	export let color = 'currentColor';
 	export let variation: 'solid' | 'outline' = 'outline';
 	let viewBox: string;
 	let svgpath: string;
-	let svgoutline = `<path d="M8.25 4.5L15.75 12L8.25 19.5" stroke="${color}" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/> `;
-	let svgsolid = `<path fill-rule="evenodd" clip-rule="evenodd" d="M16.2803 11.4697C16.5732 11.7626 16.5732 12.2374 16.2803 12.5303L8.78033 20.0303C8.48744 20.3232 8.01256 20.3232 7.71967 20.0303C7.42678 19.7374 7.42678 19.2626 7.71967 18.9697L14.6893 12L7.71967 5.03033C7.42678 4.73744 7.42678 4.26256 7.71967 3.96967C8.01256 3.67678 8.48744 3.67678 8.78033 3.96967L16.2803 11.4697Z" fill="${color}"/> `;
+	let svgoutline = `<path clip-rule="evenodd" fill-rule="evenodd" d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z" stroke="${color}"></path> `;
+	let svgsolid = `<path clip-rule="evenodd" fill-rule="evenodd" d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z" fill="${color}"></path> `;
 	$: switch (variation) {
 		case 'outline':
 			svgpath = svgoutline;
-			viewBox = '0 0 24 24';
+			viewBox = '0 0 20 20';
 			break;
 		case 'solid':
 			svgpath = svgsolid;
-			viewBox = '0 0 24 24';
+			viewBox = '0 0 20 20';
 			break;
 		default:
 			svgpath = svgoutline;
-			viewBox = '0 0 24 24';
+			viewBox = '0 0 20 20';
 	}
 	export let ariaLabel = 'chevron right';
 </script>
