@@ -1,5 +1,5 @@
 <script>
-  import 'prism-themes/themes/prism-holi-theme.css';
+  // import 'prism-themes/themes/prism-holi-theme.css';
 </script>
 
 <slot />
