@@ -1,7 +1,7 @@
 <script lang="ts">
-	export let pClass: string = 'mb-4 text-lg text-gray-500 dark:text-gray-400';
+  export let pClass: string = 'text-lg text-gray-600 dark:text-gray-400';
 </script>
 
 <p class={pClass}>
-	<slot />
+  <slot />
 </p>
